+++ Assets/Scripts/Models/ExitData.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class ExitData: TileData {
     public Action ESellFish;
 
-    override protected int ProcessingDuration{ get{return 0;}} 
+    override protected float ProcessingDuration { get { return 0; } }
     public ExitData(Direction direction, Vector2 position) : base(Appliances.Exit, position) {
 
         base.direction = direction;
